gateway-service/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
PyJWT